+++ src/resources/tts/index.ts
@@ -22,11 +22,4 @@ export {
 export { type TTSWSClientOptions } from './ws';
 export { type TTSWSReconnectOptions } from './ws-base';
 
-export {
-  TTSWS,
-  TTSWSContext,
-  WebSocketTimeoutError,
-  type ContextOptions,
-  type ContextGenerateRequest,
-} from '../../lib/tts/ws/3-0-0';
-export { TTSContexts } from '../../lib/tts/ws/context-manager';
+export { TTSWSContexts } from '../../lib/tts/ws/contexts';